Update Navbar structure
I didn't see the need for the previously used swipeable menu on the left
side, because the current frontend structure is more "tightly knit"
while previously it was more like a collection of independent screens.

- User button: pops up a window showing information about the current user
and the session. If the user is a superuser, they can also navigate to
the admin panel and the audit log from here.

- Organization: link to the CompanyHierarchy screen.

- Dashboard: link to the Dashboard screen. If previously there was a
  Station selected, the link will redirect to that Station's dashboard
  (if a Location was selected, it will open that Location). If no
  Station was selected, the user will be directed to the
  CompanyHierarchy screen.

- Data: link to the Data screen. The Data screen will contain aggregate
  views of all Detectors, Jobs, Tasks, etc. in a tabular format.

- About: link to the About screen.