Display detector stream on Stream component
Apparently it's pretty simple to display MJPEG videos: set the URL as
the source on a <img> tag (not a <video> tag! It doesn't support MJPEG)

Currently the frame rate is not tracked. Seems like the tag's onLoad
event fires whenever a new image is displayed, so we could use that.